backend/pyproject.toml
[project]
name = "shelter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "openai>=1.91.0",
    "python-dotenv>=1.1.1",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "bcrypt>=4.0.0"
]